refactor: extract selection submenus
